refactor: Separate Payment and Flow APIs (#2)
- Created `src/payments.rs` and `src/flows.rs` modules.
- Moved payment-related methods into a `Payments` struct.
- Moved flow-related methods into a `Flows` struct.
- Updated the main `Client` to provide access to the new handlers via `client.payments()` and `client.flows()`.
- Updated tests to use the new API structure.
